refactor(markdown): consolidate MDX components into markdown.tsx
Remove mdx-components.tsx and migrate its functionality into markdown.tsx for better maintainability. Also, delete unused dependencies and API route related to MDX and Fuse.js search.
