+++ components/markdown/markdown.tsx
@@ -1,11 +1,13 @@
 /* eslint-disable @next/next/no-img-element */
 import React, { useState } from 'react';
+import { ReactNode } from 'react';
 import ReactMarkdown from 'react-markdown';
 import rehypeHighlight from 'rehype-highlight';
 import rehypeSanitize from 'rehype-sanitize';
 import frontmatter from 'remark-frontmatter';
 import remarkGfm from 'remark-gfm';
 
+import CopyButton from '@/components/button/copy.button';
 import { Hidden } from '@/components/common/hidden';
 import InternalLink from '@/components/common/internal-link';
 import { Dialog, DialogContent, DialogDescription, DialogTitle, DialogTrigger } from '@/components/ui/dialog';
@@ -20,9 +22,36 @@ type MarkdownProps = {
 	children: string;
 };
 
+export const ID_REPlACE_REGEX = /[\s\\/]+/g;
+const idMaps: Record<string, number | undefined> = {};
+
+export const shared = {
+	idMaps,
+};
+
+function toId(children: ReactNode): string {
+	if (typeof children === 'string') {
+		return children;
+	}
+
+	throw new Error('Invalid heading: ' + children);
+}
+
+const Heading = ({ as: Tag, children, ...props }: { as: any; children: React.ReactNode }) => {
+	const id = toId(children).toLowerCase().replaceAll(ID_REPlACE_REGEX, '-');
+	const count = shared['idMaps'][id] ?? 0;
+	shared['idMaps'][id] = count + 1;
+
+	return (
+		<Tag id={id + '-' + count} {...props} className="scroll-mt-20">
+			{children}
+		</Tag>
+	);
+};
+
 export const OTHER_WEBSITE_URL_REGEX = /^(https?:)?\/\//;
 
-function RouterLink({ href, children }: any) {
+function A({ href, children }: any) {
 	if (href.match(YOUTUBE_VIDEO_REGEX)) {
 		const id = extractYouTubeID(href);
 		if (id) {
@@ -39,7 +68,7 @@ function RouterLink({ href, children }: any) {
 	}
 
 	return (
-		<InternalLink className="text-emerald-500" href={href}>
+		<InternalLink className="text-brand" href={href}>
 			{children}
 		</InternalLink>
 	);
@@ -58,10 +87,20 @@ function MarkdownImage({ src, alt }: any) {
 		return alt;
 	}
 
+	const width = new URL(src).searchParams.get('w') ?? undefined;
+	const height = new URL(src).searchParams.get('h') ?? undefined;
+
 	return (
 		<Dialog>
 			<DialogTrigger asChild>
-				<img className="markdown-image rounded-md max-h-[50dvh]" alt={alt} src={src} onError={() => setError(true)} />
+				<img
+					className="markdown-image rounded-md max-h-[50dvh]"
+					alt={alt}
+					src={src}
+					width={width}
+					height={height}
+					onError={() => setError(true)}
+				/>
 			</DialogTrigger>
 			<DialogContent className="max-w-[100dvw] max-h-dvh sm:max-h-dvh flex justify-center items-center h-full">
 				<Hidden>
@@ -74,13 +113,39 @@ function MarkdownImage({ src, alt }: any) {
 	);
 }
 
+const Pre = ({ className, children, ...props }: any) => (
+	<pre className={cn('relative group', className)} {...props}>
+		{children}
+	</pre>
+);
+
+const Code = ({ className, children, ...props }: any) => (
+	<>
+		<code className={cn(className)} {...props}>
+			{children}
+		</code>
+		<CopyButton className="top-1 right-1 absolute" data={children} content={children} variant="ghost" />
+	</>
+);
+
 export default function Markdown({ className, children }: MarkdownProps) {
 	return (
 		<ReactMarkdown
 			className={cn('markdown space-y-4', className)}
-			components={{ a: RouterLink, img: MarkdownImage }}
+			components={{
+				a: A,
+				img: MarkdownImage,
+				pre: Pre,
+				code: Code,
+				h1: (props: any) => <Heading as="h1" {...props} />,
+				h2: (props: any) => <Heading as="h2" {...props} />,
+				h3: (props: any) => <Heading as="h3" {...props} />,
+				h4: (props: any) => <Heading as="h4" {...props} />,
+				h5: (props: any) => <Heading as="h5" {...props} />,
+				h6: (props: any) => <Heading as="h6" {...props} />,
+			}}
 			rehypePlugins={[rehypeSanitize, rehypeHighlight]}
-      remarkPlugins={[[frontmatter,  { type: 'yaml', marker: '-' }], remarkGfm]}
+			remarkPlugins={[[frontmatter, { type: 'yaml', marker: '-' }], remarkGfm]}
 		>
 			{children}
 		</ReactMarkdown>
