+++ package.json
@@ -14,9 +14,6 @@
     },
     "dependencies": {
         "@hookform/resolvers": "^3.9.1",
-        "@mdx-js/loader": "^3.1.0",
-        "@mdx-js/react": "^3.1.0",
-        "@next/mdx": "^15.2.1",
         "@next/third-parties": "^15.1.2",
         "@radix-ui/react-accordion": "^1.2.2",
         "@radix-ui/react-alert-dialog": "^1.1.4",
@@ -59,7 +56,6 @@
         "emoji-picker-react": "^4.12.0",
         "file-saver": "^2.0.5",
         "framer-motion": "^10.18.0",
-        "fuse.js": "^7.1.0",
         "i18next": "^24.2.0",
         "i18next-browser-languagedetector": "^8.0.2",
         "i18next-chained-backend": "^4.6.2",
@@ -104,18 +100,11 @@
         "@tisoap/react-flow-smart-edge": "^3.0.0",
         "@trivago/prettier-plugin-sort-imports": "^4.3.0",
         "@types/file-saver": "^2.0.7",
-        "@types/mdx": "^2.0.13",
         "@types/negotiator": "^0.6.3",
         "@types/node": "^20.17.10",
-        "@types/nprogress": "^0.2.3",
-        "@types/pako": "^2.0.3",
         "@types/react": "^18.3.18",
         "@types/react-dom": "^18.3.5",
         "@types/react-infinite-scroller": "^1.2.5",
-        "@types/react-lazyload-youtube": "^1.0.0",
-        "@types/react-window": "^1.8.8",
-        "@types/react-window-infinite-loader": "^1.0.9",
-        "@types/remark-heading-id": "^1.0.0",
         "@types/throttle-debounce": "^5.0.2",
         "@types/web-push": "^3.6.4",
         "autoprefixer": "^10.4.20",
